+++ spec/adhearsion/dial_plan/manager_spec.rb
@@ -66,9 +66,9 @@ module Adhearsion
           }.should raise_error(Manager::NoContextError)
         end
 
-        it 'should send :accept to the execution environment if Adhearsion::AHN_CONFIG.automatically_accept_incoming_calls is set' do
+        it 'should send :accept to the execution environment if Adhearsion.config.automatically_accept_incoming_calls is set' do
           flexmock(ExecutionEnvironment).new_instances.should_receive(:accept).once.and_throw :accepted_call!
-          Configuration.configure do |config|
+          Adhearsion.config.platform do |config|
             config.automatically_accept_incoming_calls = true
           end
           the_following_code {
@@ -76,8 +76,8 @@ module Adhearsion
           }.should throw_symbol :accepted_call!
         end
 
-        it 'should NOT send :accept to the execution environment if Adhearsion::AHN_CONFIG.automatically_accept_incoming_calls is NOT set' do
-          Configuration.configure do |config|
+        it 'should NOT send :accept to the execution environment if Adhearsion.config.automatically_accept_incoming_calls is NOT set' do
+          Adhearsion.config.platform do |config|
             config.automatically_accept_incoming_calls = false
           end
 
@@ -90,8 +90,31 @@ module Adhearsion
 
           flexmock(ExecutionEnvironment).should_receive(:new).once.and_return execution_env
 
-          Configuration.configure
           subject.handle call
+        end
+      end
+
+      it "should define accessors for other contexts in the dialplan" do
+        call = new_call_for_context :am_not_for_kokoa!
+        bogus_dialplan = <<-DIALPLAN
+          am_not_for_kokoa! {}
+          icanhascheezburger? {}
+          these_context_names_do_not_really_matter {}
+        DIALPLAN
+
+        mock_dialplan_with bogus_dialplan
+
+        flexmock(call).should_receive :write_and_await_response
+
+        manager = Adhearsion::DialPlan::Manager.new
+        manager.dial_plan.entry_points.empty?.should_not be true
+
+        flexmock(call).should_receive(:hangup!).once
+
+        manager.handle call
+
+        %w(these_context_names_do_not_really_matter icanhascheezburger? am_not_for_kokoa!).each do |context_name|
+          manager.context.respond_to?(context_name).should be true
         end
       end
 
@@ -121,7 +144,7 @@ module Adhearsion
       describe "control statements" do
 
         it "should catch ControlPassingExceptions" do
-          flexmock(Adhearsion::AHN_CONFIG).should_receive(:automatically_accept_incoming_calls).and_return false
+          Adhearsion.config.platform.automatically_accept_incoming_calls = false
           dialplan = %{
             foo { raise Adhearsion::DSL::Dialplan::ControlPassingException.new(bar) }
             bar {}
@@ -178,7 +201,7 @@ module Adhearsion
         end
 
         it "new constants should still be accessible within the dialplan" do
-          flexmock(Adhearsion::AHN_CONFIG).should_receive(:automatically_accept_incoming_calls).and_return false
+          Adhearsion.config.platform.automatically_accept_incoming_calls = false
           ::Jicksta = :Jicksta
           dialplan = %{
             constant_test {