+++ spec/adhearsion/punchblock/commands/output_spec.rb
@@ -21,6 +21,7 @@ module Adhearsion
             end
 
             it 'should return false' do
+              pending
               mock_execution_environment.play_ssml(ssml).should be false
             end
           end